Put hot button and counter into HotCounter copmonent
