+++ react-hot-button-component/my-app/src/HotButton.tsx
